requirements/base.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
django-cors-headers==3.7.0
django-extensions==3.1.1
djangorestframework==3.12.2
numpy==1.20.1
pandas==1.2.1
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
